Assistant/requirements.txt
google-generativeai==0.6.0
protobuf==5.27.1
pyttsx3==2.90
SpeechRecognition==3.10.4
